# perf: reduce lock contention for internal metrics