--- a/metrics/metrics.go
+++ b/metrics/metrics.go
@@ -14,8 +14,8 @@ import (
 // objects. This allows us to support multiple metrics backends at once.
 //
 // The user can configure which metrics backends are enabled, and which are not.
-// The StoreMetrics object is always enabled, and is the first child of the
-// MultiMetrics object. This is because we always want StressRelief to be able
+// Metrics contains two more methods for retrieving metrics data: Get() and Store().
+// This is because we always want StressRelief to be able
 // to retrieve metrics data even if the user has not configured any other
 // metrics backends.
 //
@@ -33,6 +33,12 @@ import (
 // used in combination with other metrics. This is mainly to support
 // StressRelief.
 type Metrics interface {
+	MetricsBackend
+	Get(name string) (float64, bool) // for reading back a counter or a gauge
+	Store(name string, val float64)  // for storing a rarely-changing value not sent as a metric
+}
+
+type MetricsBackend interface {
 	// Register declares a metric; metricType should be one of counter, gauge, histogram, updown
 	Register(metadata Metadata)
 	Increment(name string)              // for counters
@@ -41,8 +47,6 @@ type Metrics interface {
 	Histogram(name string, obs float64) // for histogram
 	Up(name string)                     // for updown
 	Down(name string)                   // for updown
-	Get(name string) (float64, bool)    // for reading back a counter or a gauge
-	Store(name string, val float64)     // for storing a rarely-changing value not sent as a metric
 }
 
 func GetMetricsImplementation(c config.Config) *MultiMetrics {