--- a/metrics/otel_metrics.go
+++ b/metrics/otel_metrics.go
@@ -35,15 +35,12 @@ type OTelMetrics struct {
 	meter        metric.Meter
 	shutdownFunc func(ctx context.Context) error
 
-	counters   map[string]metric.Int64Counter
-	gauges     map[string]metric.Float64ObservableGauge
-	histograms map[string]metric.Float64Histogram
-	updowns    map[string]metric.Int64UpDownCounter
-
-	// values keeps a map of all the non-histogram metrics and their current value
-	// so that we can retrieve them with Get()
-	values map[string]float64
-	lock   sync.RWMutex
+	lock             sync.RWMutex
+	counters         map[string]metric.Int64Counter
+	gauges           map[string]metric.Float64Gauge
+	histograms       map[string]metric.Float64Histogram
+	updowns          map[string]metric.Int64UpDownCounter
+	observableGauges map[string]metric.Float64ObservableGauge
 }
 
 // Start initializes all metrics or resets all metrics to zero
@@ -54,11 +51,10 @@ func (o *OTelMetrics) Start() error {
 	defer o.lock.Unlock()
 
 	o.counters = make(map[string]metric.Int64Counter)
-	o.gauges = make(map[string]metric.Float64ObservableGauge)
+	o.gauges = make(map[string]metric.Float64Gauge)
 	o.histograms = make(map[string]metric.Float64Histogram)
 	o.updowns = make(map[string]metric.Int64UpDownCounter)
-
-	o.values = make(map[string]float64)
+	o.observableGauges = make(map[string]metric.Float64ObservableGauge)
 
 	ctx := context.Background()
 
@@ -155,7 +151,7 @@ func (o *OTelMetrics) Start() error {
 		return err
 	}
 
-	o.gauges[name] = g
+	o.observableGauges[name] = g
 
 	name = "memory_inuse"
 	// This is just reporting the gauge we already track under a different name.
@@ -169,7 +165,7 @@ func (o *OTelMetrics) Start() error {
 	if err != nil {
 		return err
 	}
-	o.gauges[name] = g
+	o.observableGauges[name] = g
 
 	startTime := time.Now()
 	name = "process_uptime_seconds"
@@ -181,7 +177,7 @@ func (o *OTelMetrics) Start() error {
 	if err != nil {
 		return err
 	}
-	o.gauges[name] = g
+	o.observableGauges[name] = g
 
 	return nil
 }
@@ -245,7 +241,6 @@ func (o *OTelMetrics) Increment(name string) {
 		}
 	}
 	ctr.Add(context.Background(), 1)
-	o.values[name]++
 }
 
 func (o *OTelMetrics) Gauge(name string, val float64) {
@@ -261,7 +256,6 @@ func (o *OTelMetrics) Gauge(name string, val float64) {
 			return
 		}
 	}
-	o.values[name] = val
 }
 
 func (o *OTelMetrics) Count(name string, val int64) {
@@ -277,7 +271,6 @@ func (o *OTelMetrics) Count(name string, val int64) {
 		}
 	}
 	ctr.Add(context.Background(), val)
-	o.values[name] += float64(val)
 }
 
 func (o *OTelMetrics) Histogram(name string, val float64) {
@@ -293,7 +286,6 @@ func (o *OTelMetrics) Histogram(name string, val float64) {
 		}
 	}
 	h.Record(context.Background(), val)
-	o.values[name] += val
 }
 
 func (o *OTelMetrics) Up(name string) {
@@ -309,7 +301,6 @@ func (o *OTelMetrics) Up(name string) {
 		}
 	}
 	ud.Add(context.Background(), 1)
-	o.values[name]++
 }
 
 func (o *OTelMetrics) Down(name string) {
@@ -325,22 +316,14 @@ func (o *OTelMetrics) Down(name string) {
 		}
 	}
 	ud.Add(context.Background(), -1)
-	o.values[name]--
 }
 
 func (o *OTelMetrics) Store(name string, value float64) {
-	o.lock.Lock()
-	defer o.lock.Unlock()
-
-	o.values[name] = value
+	return
 }
 
 func (o *OTelMetrics) Get(name string) (float64, bool) {
-	o.lock.RLock()
-	defer o.lock.RUnlock()
-
-	val, ok := o.values[name]
-	return val, ok
+	return 0, false
 }
 
 // initCounter initializes a new counter metric with the given metadata
@@ -364,24 +347,11 @@ func (o *OTelMetrics) initCounter(metadata Metadata) (metric.Int64Counter, error
 
 // initGauge initializes a new gauge metric with the given metadata
 // It should be used while holding the metrics lock.
-func (o *OTelMetrics) initGauge(metadata Metadata) (metric.Float64ObservableGauge, error) {
+func (o *OTelMetrics) initGauge(metadata Metadata) (metric.Float64Gauge, error) {
 	unit := string(metadata.Unit)
-	var f metric.Float64Callback = func(_ context.Context, result metric.Float64Observer) error {
-		// this callback is invoked from outside this function call, so we
-		// need to Rlock when we read the values map. We don't know how long
-		// Observe() takes, so we make a copy of the value and unlock before
-		// calling Observe.
-		o.lock.RLock()
-		v := o.values[metadata.Name]
-		o.lock.RUnlock()
-
-		result.Observe(v)
-		return nil
-	}
-	g, err := o.meter.Float64ObservableGauge(metadata.Name,
+	g, err := o.meter.Float64Gauge(metadata.Name,
 		metric.WithUnit(unit),
 		metric.WithDescription(metadata.Description),
-		metric.WithFloat64Callback(f),
 	)
 	if err != nil {
 		return nil, err