… metrics/legacy_test.go