## Which problem is this PR solving?

Refinery supports three metrics backends: libhoney (Honeycomb), OpenTelemetry (OTel), and Prometheus. Users can enable one or more of these backends simultaneously through the MultiMetrics wrapper.

Currently, each metrics implementation—including MultiMetrics—maintains its own internal locks and in-memory copies of metrics. This leads to unnecessary duplication, especially for metrics critical to the Stress Relief system, where we only need a single shared in-memory representation.

Additionally:
- OTel and Prometheus implementations are already thread-safe and do not require extra synchronization.
- The libhoney implementation previously relied on a mutex for concurrency control. This PR updates it to use atomic operations instead, reducing lock contention and improving performance under load.

## Short description of the changes

- Replace mutex with atomic operation for all metrics except histogram in libhoney metrics
- Remove inmemory metrics value in both OTel and Prometheus since MultiMetrics is storing them already
- Only hold locks for map access to retrieve a metric by name in OTel and Prometheus metrics
- Enable msgpack encoding for libhoney metrics and logger.
- Add tests and benchmark tests

## Benchmark
```
goos: darwin
goarch: arm64
pkg: github.com/honeycombio/refinery/metrics
cpu: Apple M2 Max
                                         │   old.txt    │                new.txt                │
                                         │    sec/op    │    sec/op     vs base                 │
ConcurrentAccess/ConcurrentCounters-12     437.2n ± ∞ ¹   354.0n ± ∞ ¹        ~ (p=1.000 n=1) ²
ConcurrentAccess/ConcurrentGauges-12       353.7n ± ∞ ¹   238.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
ConcurrentAccess/ConcurrentHistograms-12   482.7n ± ∞ ¹   356.8n ± ∞ ¹        ~ (p=1.000 n=1) ²
ConcurrentAccess/ConcurrentMixed-12        483.6n ± ∞ ¹   368.9n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                    435.9n         324.8n        -25.48%
```